Prevent pxe retry when agent token exists
A race condition can be observed in CI under heavy load where the
conductor triggers are boot of the agent before it is fully online
based upon state, but not considering the existence of an agent
token. As a result, agent is never able to check in with Ironic
and the overall operation fails.

We now consider agent token's existence before retrying PXE as
it is the very earliest indicator of a starting agent.

Change-Id: Ice764866a08647031d16570860ec384204269501
Story: 2010107
Task: 45674