Improve dependency resolution in driver-requirements.txt
Building the Kolla ironic-conductor image for Ubuntu using Python 3.8 in
Wallaby, I noticed the build taking a very long time (25 minutes in
Zuul vs ~2 minutes for Xena).

There is a lot of backtracking in the pip dependency resolver. I tried
limiting each of the deps in driver-requirements.txt to a minor version,
and the one that really made the difference was ansible.

Note that this does not appear to affect later branches, which happily
install Ansible 6.0.0.  It also does not affect CentOS Stream 8, which
has Python 3.6 and is therefore limited to Ansible 4.x.

Change-Id: If29a86501dcc4faa60ead056aa339031cea41cab