CI: Save routing table information for troubleshooting
Presently, we don't save the state of the routing table,
which means troubleshooting the ipv6 job failures is difficult
because we just have no way of knowing what is truly going on
with the traffic flow.

Possibly, we may be encountering a case where router
advertisements are overrriding route table entries, and
the only way to know is to see the final state of the table.

Change-Id: Ib15cd21e47e3b884bae2cb956f9113e73034391b