Fix ilo boot interface order
This change aligns the boot interface order of the ilo
hardware type to match the other hardware type interface
order lists.

This change is a result of an operator reporting inconsistent
behavior of ironic when they are adding nodes using the ilo
hardware type, where they would default to use the "pxe" boot
interface, where as the other interfaces would end up defaulting
to "ipxe".

Change-Id: I3d848af284545a7a1fb1e065f09fe2df6a9114ac