Correct Image properties lookup for paths
The image lookup process, when handed a path attempts
to issue a HEAD request against the path and gets a
response which is devoid of details like a content length
or any properties. This is expected behavior, however if
we have a path, we also know we don't need to explicitly
attempt to make an HTTP HEAD request in an attempt to
match the glance ``kernel_id`` -> ``kernel`` and similar
value population behavior.

Also removes an invalid test which was written before the
overall method was fully understood.

And fixes the default fallback for kickstart template
configuration, so that it uses a URL instead of a
direct file path.

And fix logic in the handling of image property result
set, where the code previously assumed a ``stage2``
ramdisk was always required, and based other cleanup
upon that.

Change-Id: I589e9586d1279604a743746952aeabbc483825df