Fix compatibility with oslo.db 12.1.0
oslo.db 12.1.0 has changed the default value for the 'autocommit'
parameter of 'LegacyEngineFacade' from 'True' to 'False'. This is a
necessary step to ensure compatibility with SQLAlchemy 2.0. However, we
are currently relying on the autocommit behavior and need changes to
explicitly manage sessions. Until that happens, we need to override the
default.

Change-Id: I9e095d810ff5398920e8ffd4f2f089d9b8d29335
Signed-off-by: Stephen Finucane <sfinucan@redhat.com>
(cherry picked from commit 74795abf2ff972b2b848eee98426bcdfee2a9c20)