Fix nodes stuck at cleaning on Network Service issues
Ironic validates network interface before the cleaning process,
currently invalid parameter is captured but for not others.
There is chance that a node could be stucked at the cleaning
state on networking issues or temporary service down of neutron
service.

This patch adds NetworkError to the exception hanlding to cover
such cases.

Change-Id: If20de2ad4ae4177dea10b7ebfc9a91ca6fbabdb9