CI: Fix/Update a few more jobs
Fix the postgres CI job by setting it to partition image only.

Disable the IPv6 job from running, it appears tempest is no longer
compatible with it's setup in this context and it had a higher
failure rate which kept it from being voting job, so it just doesn't
need to run now.

Set the dib build job to wholedisk only, as the image built is
a wholedisk image.

Disable ironic-tempest-ipa-partition-uefi-pxe_ipmitool, as it was
previously failing freqently due to memory consumption required.
Now it is also failing due to memory quota enforcement coming out
of nova, meaning we are likely to never fix this job on this branch.

Explicitly set the standalone job to no-bios, as the default devstack
order is fake,no-bios.

Change-Id: I6ef41d66b4c4f1a5ec532ee36fc36e98aa82655f