CI: disable bfv testing on bugfix/19.0
We appear to be getting an odd mix with Cinder, specifically with
os_brick no longer having a setup module internally.

This was a known thing that could happen on bugfix branches, and since
this is not really needing to be tested on this branch, we can disable
the CI job.

Change-Id: I124c9ca131e3aa34b700902f4984cfbdfae17df3