Fix snmp driver sleep call count
Mocking time.sleep is known to be problematic in general,
espescially when eventlet is involved. Since this can generate
false failures, change the unit test to just check that it was
called, as opposed to trying to count the number of times so
we don't accidently count other test's calling time.sleep.

Change-Id: I4323e74d7af008a651719fb972df667eb823e314