Change boot_interface order of iRMC driver
This change aligns the boot interface order of the irmc
hardware type to match the other hardware type interface
order lists.

This change is a result of an operator reporting inconsistent
behavior of ironic when they are adding nodes using the irmc
hardware type, where they would default to use the "irmc-pxe" boot
interface, where as the other interfaces would end up defaulting
to "ipxe".

Change-Id: I017c6560f9de884eefb2c1925321380cc1c721e2