Fix debug log message argument formatting
The format string is expecting a dictionary with keys matching
those used in the format string. Any unused parameters will
cause an "not all arguments converted during string formatting"
exception.

The quote style is also changed from double to single quotes to
match the other logging statements in the code.

Change-Id: Ic9dea4f51d82866be8ac16242a79237c789b9745