Catch any exception for Cleaning
No exception is used to communicate back, the exceptions
are used to catch failures, and if we don't catch other
possible exceptions leaving cleaning states, we may not clean
up state properly.

So instead of specific exceptions, we just catch any exception
like is used earlier in the same method.

Inspired by https://review.opendev.org/c/openstack/ironic/+/866856
and investigation through the code base as a result of inability
to clean the node.

Change-Id: I2a6bca3550819b98adbaffe315f77427b8a43d62