CI: Fix race prone unit test in networking
In the networking code stack, one of the methods
looks to identify if a change has occured, except
some of the other tests utilize the same value that
was previously asserted for the same base object.

Becaues of this, just use a unique value so we
don't risk the possibility of the test failing
erroneously.

Change-Id: Ide2b205ade67a4090a0b9bfe1282d01f7605ceb9