Set lockutils default logging
While developing some internal metrics collection capability,
and the realization that a lock was needed, we realized that
the lock activity itself would be a bit noisy. And image actions
also get lock logging, and it is just really noisy, but not super
helpful for troubleshooting.

So, set it to WARNING instead.

Discussion wise, see:
https://review.opendev.org/c/openstack/ironic-lib/+/865311

Change-Id: I3ab14ee5b5cc063784d26e3c760f1422c692060d