Always fall back from hard linking to copying files
The current check is insufficient: it passes for Kubernetes shared
volumes, although hard-linking between them is not possible.
This patch changes the approach to trying a hard link and falling
back to copyfile instead.

The patch relies on optimizations in Python 3.8 and thus should not
be backported beyond the Zed series to avoid performance regression.

Change-Id: I929944685b3ac61b2f63d2549198a2d8a1c8fe35