CI: Try to isolate test failures in neutron vif logic
One of our tests is super-friendly to race because it relies on changing
an object, but that object not being saved yet to verify the difference.
However, that object and even similar changes can occur in the database
and trigger the test to fail.

So instead using the shared resource created on test startup, create an
isolated, dedicated resource for it to use inside of the test, which should
eliminate the entire issue. I hope.

Change-Id: Ia0df1c211d5510777c6055f1341335f9bc5c5006