CI: Change tinycore URL
We'v been able to observe one of the scenario test jobs failing
due to tinycorelinux being inaccessible. Possibly on an IPv6 only
test VM. Turns out tinycorelinux's main page is only accessible via
IPv4.

As such, I've changed the mirror to a mirror which is acessible via
IPv6 and which I've verified works for me.

Change-Id: I2b4ccd16189038ce2f054d7403775b012796aea3