CI: DB: Don't return inside of node get wrappers
Previously, we did query object handlers with model_query
where we woudl then return the handler, however that was problematic
because we returned ORM objects.

Now we have selects, but the data has to be extrated to a set before
we close out the reader session with the database. Returning the
overall object means that the call seems to sometimes not completely
execute on newer python versions in unit tests, which create unit
test failures.

We now hold off the database query and operation to a variable,
exit the reader, and return the variable.

Change-Id: I6ae9e4442bcb473ab5ccff6e167bf61f3daa0f7e