CI: Change migrations timeout to be >60 seoncds
In local testing, I found the migrations tended to take an
average of 70 seconds. Granted, my test machine is old, and slow
but the performance is very similar to a busy cloud provider.

As such, increase the timeout to a larger value so we can enable
the double migration test again.

Also use BASE_TEST_TIMEOUT as time limit for unit tests, failing
hard if that's passed.

Change-Id: Ida495d59bde977a20590c34e282c884a65ceab43