Fix the HTTP code of the BadRequest exception
We have both Invalid and BadRequest which result in HTTP 400 and 500
accordingly. The latter is clearly incorrect.

Then we have NotFound and HTTPNotFound which mean the same thing.

Alias exceptions in both cases with the intention to drop one copy.

Finally, NotAuthorized and Unauthorized result in HTTP 403 and 500
again. Fortunately, the latter is not used and can be removed.

Change-Id: If9d571792a8617dd6ecf17e163dea252cb0f7fae